requirements.txt
beautifulsoup4==4.6.3
numpy==1.17.3
pandas==0.23.4
pytube3==9.6.4
moviepy==1.0.3
tqdm==4.46.0